requirements.txt
gdown
einops
mmengine
timm
SimpleITK
itk-elastix
pyncclient
ml_collections
psutil
scikit-image
torchinfo
antspyx
py-cpuinfo